# docs/api/phase1-auth.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThunderForgeVTT Phase 1 Authentication and User Data API",
    "version": "v1",
    "description": "Public contract for local auth, OAuth completion/linking, session lifecycle, self-service data export, and permanent deletion."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment"
    }
  ],
  "paths": {
...
    "/user/data/export": {
      "get": {
        "summary": "Export the authenticated user's data",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["json", "zip"],
              "default": "json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON or ZIP export payload"
          }
        }
      }
    },
...
  },
...
}